Convert into two docs parsing backends. Make doc parsing backend configurable.
