+++ antsibull.cfg
@@ -5,6 +5,7 @@ process_max = none
 pypi_url = https://pypi.org/
 thread_max = 80
 max_retries = 10
+doc_parsing_backend = ansible-doc
 logging_cfg = {
     version = 1.0
     outputs = {
